+++ package.json
@@ -102,7 +102,8 @@
           "src/scripts/dfnshelper.js",
           "src/scripts/imagehelper.js",
           "src/scripts/inputManager.js",
-          "src/components/deleteHelper.js"
+          "src/components/deleteHelper.js",
+          "src/components/actionSheet/actionSheet.js"
         ],
         "plugins": [
           "@babel/plugin-transform-modules-amd"
